refactor : move calculateDistance to HelperService
