+++ common/helper/helper.service.ts
@@ -2,12 +2,22 @@ import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class HelperService {
-  calculateDistance() // lat1: number,
-  // lon1: number,
-  // lat2: number,
-  // lon2: number,
-  : number {
-    // Harversine formula goes here
-    return 0;
+  calculateDistance(
+    lat1: number,
+    lon1: number,
+    lat2: number,
+    lon2: number,
+  ): number {
+    const toRad = (value: number) => (value * Math.PI) / 180;
+    const R = 6371; // Earth radius in km
+
+    const dLat = toRad(lat2 - lat1);
+    const dLon = toRad(lon2 - lon1);
+    const a =
+      Math.sin(dLat / 2) ** 2 +
+      Math.cos(toRad(lat1)) * Math.cos(toRad(lat2)) * Math.sin(dLon / 2) ** 2;
+
+    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+    return R * c;
   }
 }
